Agent_with_phi/requirements.txt
python-dotenv
phidata
yfinance
packaging
duckduckgo-search
fastapi 
uvicorn
groq